+++ TIMEO/src/main/java/it/nextworks/nfvmano/timeo/sbdriver/sdn/tapi/TapiSupportedWavelength.java
@@ -2,21 +2,21 @@ package it.nextworks.nfvmano.timeo.sbdriver.sdn.tapi;
 
 public class TapiSupportedWavelength {
 
-    private long upperFrequency;
-    private long lowerFrequency;
+    private float upperFrequency;
+    private float lowerFrequency;
     private TapiFrequencyConstraint tapiFrequencyConstraint;
 
-    public TapiSupportedWavelength(long upperFrequency, long lowerFrequency, TapiFrequencyConstraint tapiFrequencyConstraint) {
+    public TapiSupportedWavelength(float upperFrequency, float lowerFrequency, TapiFrequencyConstraint tapiFrequencyConstraint) {
         this.upperFrequency = upperFrequency;
         this.lowerFrequency = lowerFrequency;
         this.tapiFrequencyConstraint = tapiFrequencyConstraint;
     }
 
-    public long getUpperFrequency() {
+    public float getUpperFrequency() {
         return upperFrequency;
     }
 
-    public long getLowerFrequency() {
+    public float getLowerFrequency() {
         return lowerFrequency;
     }
